+++ src/pybend/core/main.py
@@ -2,6 +2,10 @@
 
 import os
 import config
+import authorize
+
+authorize.configure(jwt_secret=config.JWT_SECRET, jwt_expiry_hours=config.JWT_EXPIRY_HOURS)
+
 from api.routes_fastapi import register_route
 from models.comment_model import Comment
 from models.like_model import Like
@@ -19,7 +23,6 @@ storage_backend = SQLiteStorage(config.SQLITE_DB_FILE)
 register_model(Product, storage=storage_backend)
 register_model(User, storage=storage_backend)
 register_model(generate_join_model(Product, Comment), storage=storage_backend)
-register_model(generate_join_model(Comment, Comment), storage=storage_backend)   # CommentComment (replies)
 register_model(generate_join_model(Comment, Like), storage=storage_backend)      # CommentLike (likes)
 
 if os.getenv("GENERATE_DOCS", "true").lower() in ("1", "true"):